// skybox/types.ts
export type GradientLayerShaderBinding = {
  index: number;
  layer: Extract<SkyboxManifestLayer, { type: "gradient" }>;
  parameterPrefix: string;
  stopCount: number;
};
// ...
export type GradientUniformNodes = {
  axis: ReturnType<typeof uniform>;
  layerId: string;
  stops: Array<{
    color: ReturnType<typeof uniform>;
    midpoint: ReturnType<typeof uniform>;
    t: ReturnType<typeof uniform>;
  }>;
};
// ...
export type BuiltInWebGpuLayerAdapter<
  TType extends SkyboxManifestLayer["type"],
  TBinding,
  TUniforms,
> = WebGpuLayerAdapter<Extract<SkyboxManifestLayer, { type: TType }>, TBinding, TUniforms>;